Changed OpenMathTerminal Structure, added set vector and complex (unfinished)
Added new abstract Wrapper-Classes for Numerics (Simple Numerics) and
Text. These abstract Wrapper-Classes are needed for OpenMathSet,
OpenMathVector and OpenMathComplex. Sets can contain all of them, but
Vector do not contain sets (but everything else) and Complex Numbers do
not contain Vectors, but all simple Numbers(variables, float, integer).

Because Of Generics, it is not possible to select, whether this Key is
an instance of X OR an instance of Y. This leads to this hierarchy.