Added Dependency from JAXBOpenMath-2.0.jar from Gitlab
Currently Windows is not supported, but soon it will be.

INFO: Configure the .sh script (add your private Token from gitlab in it
(in YOURTOKEN="add it here")) and run this script once to initialize it
for maven.

Same goes for Windows users with the corresponding .bat file.

Now by generate-sources, build, install and all other options the newest
version of JAXBOpenMath will be downloaded automatically.